# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Chat Gateway API — Qwen AI proxy with API key management
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  code?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ChatRequest {
  prompt: string;
  model?: string;
  conversationId?: string | null;
  apiKey?: string | null;
}

export interface ChatResponse {
  response: string;
  thinking?: string | null;
  model: string;
  conversationId: string;
  tokenUsage?: number | null;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
  thinking?: string | null;
  timestamp: string;
}

export interface ChatSession {
  conversationId: string;
  model: string;
  messages: ChatMessage[];
  createdAt: string;
  updatedAt: string;
  messageCount: number;
}

export interface ChatHistoryList {
  sessions: ChatSession[];
  total: number;
  offset: number;
  limit: number;
}

export interface GatewayStats {
  totalRequests: number;
  successRequests: number;
  failedRequests: number;
  successRate: number;
  requestsToday: number;
  requestsThisHour: number;
  averageResponseTime: number;
  activeApiKeys: number;
  tokenPoolSize: number;
  tokenPoolHealthy: number;
}

export interface ApiKey {
  id: string;
  name: string;
  /** Masked key (e.g. gw_****abcd) */
  keyPreview: string;
  createdAt: string;
  lastUsed?: string | null;
  requestCount: number;
  isActive: boolean;
}

export interface ApiKeyCreated {
  id: string;
  name: string;
  /** Full API key — only shown once */
  key: string;
  keyPreview: string;
  createdAt: string;
  isActive: boolean;
}

export interface ApiKeyList {
  keys: ApiKey[];
  total: number;
}

export interface CreateApiKeyRequest {
  name: string;
}

export interface UpdateApiKeyRequest {
  isActive?: boolean;
  name?: string;
}

export interface QwenModel {
  id: string;
  name: string;
  description: string;
  maxTokens: number;
}

export interface ModelList {
  models: QwenModel[];
}

export interface TokenPoolStatus {
  total: number;
  healthy: number;
  exhausted: number;
  lastRefreshed?: string | null;
  nextRefresh?: string | null;
  rotationCount: number;
}

export type GetChatHistoryParams = {
/**
 * @minimum 1
 * @maximum 200
 */
limit?: number;
/**
 * @minimum 0
 */
offset?: number;
};