# lib/api-zod/src/generated/types/apiKey.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Chat Gateway API — Qwen AI proxy with API key management
 * OpenAPI spec version: 0.1.0
 */

export interface ApiKey {
  id: string;
  name: string;
  /** Masked key (e.g. gw_****abcd) */
  keyPreview: string;
  createdAt: string;
  lastUsed?: string | null;
  requestCount: number;
  isActive: boolean;
}