# lib/api-zod/src/generated/types/apiKeyList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Chat Gateway API — Qwen AI proxy with API key management
 * OpenAPI spec version: 0.1.0
 */
import type { ApiKey } from './apiKey';

export interface ApiKeyList {
  keys: ApiKey[];
  total: number;
}