# lib/api-zod/src/generated/types/tokenPoolStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Chat Gateway API — Qwen AI proxy with API key management
 * OpenAPI spec version: 0.1.0
 */

export interface TokenPoolStatus {
  total: number;
  healthy: number;
  exhausted: number;
  lastRefreshed?: string | null;
  nextRefresh?: string | null;
  rotationCount: number;
}